backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
web3==6.11.3
python-dotenv==1.0.1
py-solc-x==1.1.1
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
